requirements.txt
openpyxl==3.0.9
xlrd==2.0.1
pandas_profiling==3.1.0
statsmodels==0.13.2
scikit-learn==0.24
lightgbm==3.3.2
xgboost==1.5.2
hyperopt==0.2.7
pyqt5==5.15.6